backend/requirements.txt
# Core scraping
selenium==4.20.0
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0

# Scheduling
schedule==1.2.2

# MongoDB Atlas support
pymongo==4.6.1
python-dotenv==1.0.0

# WhatsApp (optional — uncomment if you use WhatsApp notifications)
# twilio==8.13.0

# Note: SQLite is part of Python's standard library — no install needed.
